Move lattice fucntions to clkLattice.cpp
